backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
reportlab
pymupdf==1.24.9
python-docx==1.1.2

scikit-learn==1.5.1
numpy==2.0.1